Update server API
Server API is updated. Web clients will use the prefix ‘webclient’ and
application clients will user ‘appclient’. Thymeleaf dependency
included to handle MVC pattern. Controller is refactored. Multiple
controllers created depending on actions to be done.
