+++ src/main/resources/static/css/login.css
@@ -1,0 +1,3 @@
+.vertical-offset-100{
+    padding-top:100px;
+}
