+++ src/main/resources/templates/to_do_template.html
@@ -1,0 +1,6 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+	<body>
+		<p>This is the default template. This method is to be implemented.</p>
+	</body>
+</html>
